requirements.txt
streamlit
PyMuPDF
faiss-cpu
spacy
sentence-transformers
llama-cpp-python
